% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_manipulation_functions.r
\name{create_rRNA}
\alias{create_rRNA}
\title{Function to create a rRNA object}
\usage{
create_rRNA(
  id,
  symbol = NA_character_,
  name = NA_character_,
  description = NA_character_,
  assembly,
  chromosome,
  start,
  end,
  strand,
  sequence = Biostrings::DNAString(""),
  transcript_id = NA_character_,
  type = "rRNA",
  RNA_sequence = Biostrings::RNAString(""),
  rRNA_type,
  ribosomal_subunit = NA_character_
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Description of the gene.}

\item{assembly}{Assembly used to store the position information.}

\item{chromosome}{Chromosome on which the gene is located.}

\item{start}{Start position of the gene on the chromosome (1-based).}

\item{end}{End position of the gene on the chromosome.}

\item{strand}{Strand of the gene.}

\item{sequence}{DNA sequence of the gene}

\item{transcript_id}{ID of the alternative transcript considered (Ex: Ensembl
Transcript ID or RefSeq ID)}

\item{type}{Type of the non-coding RNA ('lncRNA', 'siRNA','rRNA', 'tRNA'...)}

\item{RNA_sequence}{RNA sequence.}

\item{rRNA_type}{Type of rRNA subunit. The ones accepted are "18S", "28S",
"5S", "5.8S".}

\item{ribosomal_subunit}{In which subunit of the ribosome the rRNA is found.}
}
\value{
An object of class rRNA
}
\description{
Creates an object of class rRNA with the specified parameters
}
