% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_manipulation_functions.r
\name{create_snRNA}
\alias{create_snRNA}
\title{Function to create a snRNA object}
\usage{
create_snRNA(
  id,
  symbol = NA_character_,
  name = NA_character_,
  description = NA_character_,
  assembly,
  chromosome,
  start,
  end,
  strand,
  sequence = Biostrings::DNAString(""),
  transcript_id = NA_character_,
  type = "snRNA",
  RNA_sequence = Biostrings::RNAString(""),
  snRNA_class,
  subclass = NA_character_,
  associated_proteins = character(0)
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Description of the gene.}

\item{assembly}{Assembly used to store the position information.}

\item{chromosome}{Chromosome on which the gene is located.}

\item{start}{Start position of the gene on the chromosome (1-based).}

\item{end}{End position of the gene on the chromosome.}

\item{strand}{Strand of the gene.}

\item{sequence}{DNA sequence of the gene}

\item{transcript_id}{ID of the alternative transcript considered (Ex: Ensembl
Transcript ID or RefSeq ID)}

\item{type}{Type of the non-coding RNA ('lncRNA', 'siRNA','rRNA', 'tRNA'...)}

\item{RNA_sequence}{RNA sequence.}

\item{snRNA_class}{A short string that defines a specific type of snRNA,
usually differentiated by function. (Ex: U1, U2...)}

\item{subclass}{Subclass of the snRNA, defined by its localisation: either
'snoRNA' or 'scaRNA'.}

\item{associated_proteins}{A list of the ID of the proteins that the snRNA
interacts with.}
}
\value{
An object of class snRNA
}
\description{
Creates an object of class snRNA with the specified parameters
}
\examples{
# ncRNA_snRNA <- create_snRNA(
# id = "ENSG00000206652",
# symbol = "RNU1-1",
# name = "RNA, U1 small nuclear 1",
# description = "",
# assembly = "hg38",
# chromosome = "chr1",
# start = 16514122,
# end = 16514285,
# strand = "-",
# sequence = "tttcatacttacctggcagg",
# RNA_sequence = 'aaaguaug',
# transcript_id = "ENST00000383925.1",
# snRNA_class = 'U1')

}
