% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_manipulation_functions.r
\name{create_tRNA}
\alias{create_tRNA}
\title{Function to create a tRNA object}
\usage{
create_tRNA(
  id,
  symbol = NA_character_,
  name = NA_character_,
  description = NA_character_,
  assembly,
  chromosome,
  start,
  end,
  strand,
  sequence = Biostrings::DNAString(""),
  transcript_id = NA_character_,
  type = "tRNA",
  RNA_sequence = Biostrings::RNAString(""),
  aminoacid,
  anticodon = NA_character_
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Description of the gene.}

\item{assembly}{Assembly used to store the position information.}

\item{chromosome}{Chromosome on which the gene is located.}

\item{start}{Start position of the gene on the chromosome (1-based).}

\item{end}{End position of the gene on the chromosome.}

\item{strand}{Strand of the gene.}

\item{sequence}{DNA sequence of the gene}

\item{transcript_id}{ID of the alternative transcript considered (Ex: Ensembl
Transcript ID or RefSeq ID)}

\item{type}{Type of the non-coding RNA ('lncRNA', 'siRNA','rRNA', 'tRNA'...)}

\item{RNA_sequence}{RNA sequence.}

\item{aminoacid}{Name of the aminoacid transported by the tRNA (can be either
full name, or 1-3 letters code notation).}

\item{anticodon}{Anticodon contained in the tRNA (rom 5' to 3').}
}
\value{
An object of class tRNA
}
\description{
Creates an object of class tRNA with the specified parameters
}
\examples{
# ncRNA_tRNA_Phe <- create_tRNA(
# id = "ENSG00000210049.1",
# name = "Mitochondrial tRNA phenylalanine",
# assembly = "hg38",
# chromosome = "chrM",
# start = 577,
# end = 647,
# strand = "+",
# transcript_id = "ENST00000387314.1",
# type = "tRNA",
# aminoacid = "phenylalanine",
# anticodon = "GAA")
}
