% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_manipulation_functions.r
\name{createProteinCodingGene}
\alias{createProteinCodingGene}
\title{Create a Protein Coding Gene object}
\usage{
createProteinCodingGene(
  id,
  symbol = NA_character_,
  name = NA_character_,
  description = NA_character_,
  assembly,
  chromosome,
  start,
  end,
  strand,
  sequence = Biostrings::DNAString(""),
  transcript_id = NA_character_,
  exons,
  protein_id,
  protein_name = NA_character_,
  protein_sequence = Biostrings::AAString("")
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Description of the gene.}

\item{assembly}{Assembly used to store the position information.}

\item{chromosome}{Chromosome on which the gene is located.}

\item{start}{Start position of the gene on the chromosome (1-based).}

\item{end}{End position of the gene on the chromosome.}

\item{strand}{Strand of the gene.}

\item{sequence}{DNA sequence of the gene}

\item{transcript_id}{ID of the alternative transcript considered
(Ex: Ensembl Transcript ID or RefSeq ID)}

\item{exons}{GRanges object representing the exons of the gene.}

\item{protein_id}{Protein ID associated with the gene (preferably from
UniProtKB).}

\item{protein_name}{Protein name (preferably from UniprotKB).}

\item{protein_sequence}{The aminoacids sequence.}
}
\value{
A ProteinCodingGene object.
}
\description{
This function creates an instance of the ProteinCodingGene class.
}
\examples{
# gene <- createProteinCodingGene(
# id = "ENSG00000139618.19",
# symbol = "BRCA2",
# name = "BRCA2 DNA repair associated",
# description = "Plays a role in DNA repair...",
# assembly = "hg38",
# chromosome = "chr13",
# start = 32315508,
# end = 32400268,
# strand = "+",
# sequence = "ACGTACCGTCAAA...",
# transcirpt_id = "ENST00000544455.6",
# exons = exons,                    # A Granges() object defined previously
# protein_id = "P51587",
# protein_name = "Breast cancer type 2 susceptibility protein",
# protein_sequence = "mpigskerptffeifktrcn....")
}
