% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_manipulation_functions.r
\name{get_slot}
\alias{get_slot}
\title{Get a slot value}
\usage{
get_slot(object, slot_name)
}
\arguments{
\item{object}{Object of Gene, and its derivates, class.}

\item{slot_name}{Name of the slot.}
}
\value{
Slot value
}
\description{
Get a slot value without directly having to access it.
}
\examples{
# Access the mature sequence of a miRNA
get_slot(my_object, 'mature_sequence')
}
