% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_package_functions.r
\name{importGeneFromEnsembl}
\alias{importGeneFromEnsembl}
\title{Import a Gene from ENSEMBL}
\usage{
importGeneFromEnsembl(ensembl_id, mart = NULL)
}
\arguments{
\item{ensembl_id}{Gene ID, specifically the ENSEMBL one.}

\item{mart}{An object of type Mart from package biomaRT. It establishes a
connection with the biomaRT database to interact with ENSEMBL data.}
}
\value{
An object of class Gene.
}
\description{
This function imports gene information from ENSEMBL (assembly: GRCh38) and
returns an object of class Gene.
}
