% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_definition.r
\docType{class}
\name{miRNA-class}
\alias{miRNA-class}
\title{Derived class: miRNA}
\description{
Class to define microRNA
}
\section{Slots}{

\describe{
\item{\code{target_genes}}{vector of Ensembl ID of the gene target of this microRNA.}

\item{\code{seed}}{Eight-base seed region of the miRNA.}

\item{\code{mature_sequence}}{Mature miRNA.}
}}

\examples{
# Seed region
# The binding between miRNAs and target genes is not perfect across the whole
# mature miRNA sequence: in mammals it is dominated by the so-called seed
# region.
# This seed region at the 5’ end of the mature miRNA consists of eight
# nucleotides.

# Mature sequence
# The generation of miRNAs is a multistage process. Briefly, the mature ∼22 nt
# miRNA sequence is embedded in one strand of an ∼33 bp double-stranded stem
# characteristic of hairpin structures
# in primary miRNA (pri-miRNA) transcripts produced by RNA polymerase II or
# III. The miRNA must therefore be excised during its biogenesis to elicit gene
#  silencing. To quickly find the mature sequence of the miRNA, the user can
#  check this database: https://mirbase.org/.
}
