% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_definition.r
\docType{class}
\name{ncRNA-class}
\alias{ncRNA-class}
\title{Derived class: ncRNA (Non Coding RNA)}
\description{
A class to define non-coding RNA.
}
\section{Slots}{

\describe{
\item{\code{transcript_id}}{Transcript ID.}

\item{\code{type}}{Non coding RNA can be: 'lncRNA' (long non-coding RNA), 'siRNA'
(small interfering RNA) and, these ones of which there are also specific
classes available, 'miRNA' (microRNA), 'snRNA' (small nuclear RNA),
'tRNA' (transfer RNA), 'rRNA' (ribosomal RNA)}

\item{\code{RNA_sequence}}{RNA sequence}
}}

