% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_definition.r
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\title{Derived gene class: ProteinCodingGene}
\description{
This is a derived class specifically for protein coding genes.
}
\section{Slots}{

\describe{
\item{\code{transcript_id}}{Transcript ID (either from Ensembl or RefSeq).}

\item{\code{exons}}{GRanges object with exons positions.}

\item{\code{protein_id}}{ID of the protein (best if taken from UniProt).}

\item{\code{protein_name}}{Name of the protein (best if taken from UniProt).}

\item{\code{protein_sequence}}{Aminoacid sequence.}
}}

\examples{
# Example of creating a GRanges object for exons
# library(GenomicRanges)
# exons <- GRanges(
#   seqnames = Rle(c("chr13")),
#   ranges = IRanges(start = c(32315507,32316421,32319076,32325075,32326100,
#   32326241,32326498,32329442,32330918,32332271,32336264,32344557,32346826,
#   32354860,32356427,32357741,32362522,32363178,32370401,32370955,32376669,
#   32379316,32379749,32380006,32394688,32396897,32398161),
#                    end = c(32315667,32316527,32319325,32325184,32326150,
#                    32326282,32326613,32329492,32331030,32333387,32341196,
#                    32344653,32346896,32355288,32356609,32357929,32362693,
#                    32363533,32370557,32371100,32376791,32379515,32379913,
#                    32380145,32394933,32397044,32400268)),
#   strand = Rle(c("+"))
# )

}
