% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_package_functions.r
\name{readFasta}
\alias{readFasta}
\title{Convert FASTA file.}
\usage{
readFasta(filepath)
}
\arguments{
\item{filepath}{The FASTA file path.}
}
\value{
A Biostrings::DNAString that contains your DNA sequence.
}
\description{
Reads a FASTA file and returns a Biostring::DNAString.
}
\examples{
# sequence <- readFasta('C:/User/Desktop/file.fa')
}
