% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_definition.r
\docType{class}
\name{rRNA-class}
\alias{rRNA-class}
\title{Class rRNA}
\description{
A class that represents genes that code for ribosomal rRNA (rRNA)
}
\section{Slots}{

\describe{
\item{\code{rRNA_type}}{Type of rRNA subunit. The ones accepted are "18S", "28S",
"5S", "5.8S".}

\item{\code{ribosomal_subunit}}{In which subunit of the ribosome the rRNA is found.}
}}

\examples{
# Ribosomal subunit
# A ribosome is composed by a small and a big subunit. For prokaryotes the
# small subunit is denoted as 30S and the big one as 50S, while for eukaryotes
# the former is 40S and the latter is 60S.
}
