% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_definition.r
\docType{class}
\name{snRNA-class}
\alias{snRNA-class}
\title{Class snRNA}
\description{
A class that represents genes that code for small nuclear RNA (snRNA)
}
\section{Slots}{

\describe{
\item{\code{snRNA_class}}{A short string that defines a specific type of snRNA,
usually differentiated by function. (Ex: U1, U2...)}

\item{\code{subclass}}{Subclass of the snRNA, defined by its localisation: either
'snoRNA' or 'scaRNA'.}

\item{\code{associated_proteins}}{A list of the ID of the proteins that the snRNA
interacts with.}
}}

\examples{
# Subclasses
By definition, snRNAs are small molecules of RNA that can be found in the
nucleus. Two subclasses can be identified depending on the specific
localization inside the nucleus:
- snoRNA, is the snRNA that if located in the nucleous;
- scaRNA, is the snRNA located in the Cajal body.
}
