% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_package_functions.r
\name{validateGeneInputs}
\alias{validateGeneInputs}
\title{Validate sequence and chromosome coordinates}
\usage{
validateGeneInputs(start, end, strand, sequence = Biostrings::DNAString(""))
}
\arguments{
\item{start}{The position at which the gene starts (1 based).}

\item{end}{The position at which the gene ends.}

\item{strand}{The DNA strand where the gene is located.}

\item{sequence}{The DNA sequence (given as string, DNAString object, FASTA
file or path containing a FASTA file.)}
}
\value{
A DNAString object
}
\description{
This function validates the input sequence and chromosome coordinates.
The sequence is returned as a DNAString (from Biostrings) if it complies
with the conditions.
}
\examples{
# validateGeneInputs(123, 344, '+', 'my_fasta_file.fa')

}
