% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_manipulation_functions.r
\name{create_miRNA}
\alias{create_miRNA}
\title{Create a miRNA Gene object}
\usage{
create_miRNA(
  id,
  symbol = NA_character_,
  name = NA_character_,
  description = NA_character_,
  assembly,
  chromosome,
  start,
  end,
  strand,
  sequence = Biostrings::DNAString(""),
  transcript_id = NA_character_,
  type = "miRNA",
  RNA_sequence = Biostrings::RNAString(""),
  target_genes,
  seed = NA_character_,
  mature_sequence = Biostrings::RNAString("")
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Description of the gene.}

\item{assembly}{Assembly used to store the position information.}

\item{chromosome}{Chromosome on which the gene is located.}

\item{start}{Start position of the gene on the chromosome (1-based).}

\item{end}{End position of the gene on the chromosome.}

\item{strand}{Strand of the gene.}

\item{sequence}{DNA sequence of the gene}

\item{transcript_id}{ID of the alternative transcript considered (Ex: Ensembl
Transcript ID or RefSeq ID)}

\item{type}{Type of the non-coding RNA ('lncRNA', 'siRNA','rRNA', 'tRNA'...)}

\item{RNA_sequence}{RNA sequence.}

\item{target_genes}{A list containing the IDs of the target genes of the
microRNA.}

\item{seed}{Seed sequence (usually around 8 nucleotides).}

\item{mature_sequence}{Mature sequence of the miRNA (usally around 22
nucleotides).}
}
\value{
A miRNA object.
}
\description{
This function creates an instance of the MicroRNA class.
}
\examples{
# miRNA125a <- create_miRNA(
# id = "ENSG00000208008",
# symbol = "MIR125A",
# name = "microRNA 125a",
# description = "MIR125A (MicroRNA 125a) is an RNA Gene, and is affiliated
# with the miRNA class. Diseases associated with MIR125A include
# Medulloblastoma and Lung Cancer.",
# assembly = "hg38",
# chromosome = "chr19",
# start = 51693254,
# end = 51693339,
# strand = "+",
# sequence = "ACGTACCGTCAAA",
# transcript_id = "ENST00000385273.1",
# type = 'miRNA',
# RNA_sequence = "UGCCAGUCUC",
# target_genes = c("ENSG00000141736", "ENSG00000065361"),
# seed = 'GAGUCCC',
# mature_sequence = "ucccugagacccuuuaaccuguga")

}
